Added initial setup of the DFA class
Changed handling of ACCEPTing states: the ACCEPT pseudo character in State is
now used as a flag, allowing us to have the DFA construction determine where
the accept states are. Also, at that point the inheritance status of the
accept state can be determined.

Strings are no longer defined as a separate pseudo character type as they
eventually have to be converted to a series of states anyway.

Following the state (and rule) construction character classes containing
sequences of subsequent characters are determined and all
character-transitions are replaced by character range transitions. This should
normally dramatically reduce the number of characters to inspect when de DFA
is constructed, and it also allows us to reduce the final column dimention of
the DFAs.

Next step: DFA construction