Added (public) std::string const &filename, (protected) void setFilename__, and  static std::string istreamName__() to the flexc++.base.h skeleton.
filename returns the name of the currently processed file.
setFilename__ changes the name of the currently processed file, and
istreamName__ returns the next name for an unnamed istream, which is <istream

The switchStreams and pushStream__(istream) members use istreamName__ to
set a name for the istream. Calling functions may use setFilename to reset the
name to a more appropriate one. popStream__ restores the name of the most
recently pushed stream.