This is a major WIP update. Preparing BOL statuses for rules, parser/rules/rulesection now defines a rule recognizing BOL lines, and parser/rules/pattern doesn't anymore.
Lots of changes related to handling DFAs, defining the new class AccCount
(accept-count) which is only used for computing accept-counts with rules using
the LOP. The technical documentation needs to be modified as well, and
AccCount computations is WIP, eventually resulting in the removal of old
functions and in the definition of new functions, now available in concept or
partially implemented.

The current sources properly compile, but do not yet result in a program
otherwise doing something useful.