Bootstrap-run completed.
The lex() function created by the current flexc++ program was used to create a
new flexc++ program (sources in the directory scanner.new). This new scanner
again created a lex function, which was used to create once again a lex
function using the scanner's specification file `lexer'. The files created by
the two latest versions of flexc++ do not differ (ignoring time-stamp
differences and #line directive numbers).

This commit still has the old scanner code under ./scanner, but the next one
will remove that directory and will use a hyperlink 'scanner' pointing to
./scanner.new.

The next step is to work towards flexc++ that uses the new scanner-code