JP encountered a segfault with the following lexer file:
----------------------------------------------------
%%
.                       std::cout << "foo\n";
.+/\)           std::cout << "anything\n";
----------------------------------------------------

This problem is presumbly solved after this commit. In the tech manual
(section 0.1.13.1: The logic used in 'setTailCount') all 16 possible
combinations of d_sawAccept and the PRE ACCEPT POST flags are covered,
describing what to do for each of these combinations. At the end of the table
several 'no action' alternatives are listed. At these alternatives the action
to perform is in fact 'no action', after which the tailCount's flag should be
set to PROCESSED. This statement was lacking, causing an infinite recursion,
eventually resulting in a segfault, for JP's lexer file.

This modification also bumps flex's version to 0.8.3