Reflex OK. Removed the additional "'"-rule from scanner/lexer
The additional rule

        "'"             return d_inCharClass ? Parser::STRING : '\'';

was added at commit 2c401f9a1c4ef83ac5c7e15e2741caf24e8cb31d. At that time it
was unclear why the rule was needed. Now it is clear that this was required
because of a faulty interpretation of matched rules: when no further
transitions are possible the scanner is not necessarily in a state in which a
rule is matched. But before that rules may have been matched.

The skeleton's member updateFinals__, called by lex__ at each read character,
updates d_final.std and d_final.bol, so they contain the latest rule and
match-length at the scanner's current state. When continuation isn't possible
anymore d_final contains information about the latest matched rules and
lengths, separately for std and BOL rules.